TC#61 I have refactored the code and fixed errors with Image and also with server and client components
