+++ src/app/page.tsx
@@ -1,6 +1,7 @@
-import { Container, Partners } from '@/shared/components/shared';
-import ServerCarousel from '@/shared/components/shared/carousel/server-carousel';
+import { Carousel, Container, Partners } from '@/shared/components/shared';
 import { MainTitles } from '@/shared/components/shared/main-titles/main-titles';
+
+import React from 'react';
 
 export default async function Page() {
   return (
@@ -8,7 +9,7 @@ export default async function Page() {
       <Container>
         <MainTitles title={'Кыргызстанское сообщество любителей тенниса'} subtitle={'Искусство становиться первым!'} />
       </Container>
-      <ServerCarousel />
+      <Carousel />
       <Partners />
     </>
   );
