+++ src/shared/components/shared/carousel/carousel.module.css
@@ -1,13 +1,5 @@
-.mainTextInBlockCarousel {
-    @apply text-[20px] font-bold uppercase mb-[25px] md:text-[64px] md:mb-[16px]
-}
-
-.underTextInBlockCarousel {
-    @apply hidden text-[#808080] md:block text-[36px] font-medium pb-[56px]
-}
-
 .carouselContainer {
-    @apply px-4 lg:px-[50px] mb-5 overflow-hidden
+    @apply px-4 mb-[80px] lg:mb-[120px] lg:px-[50px] overflow-hidden
 }
 
 .carouselContent {
@@ -19,5 +11,5 @@
 }
 
 .carouselFile {
-    @apply w-full h-[244px] sm:h-[400px] md:h-[450px] lg:h-[662px] object-cover rounded-lg
+    @apply w-full h-[244px] sm:h-[400px] md:h-[450px] lg:h-[662px] xl:h-[900px] object-cover rounded-lg
 }
