--- src/shared/components/shared/carousel/server-carousel.tsx
@@ -1,12 +1,0 @@
-import { getCarousel } from '@/actions/carousel';
-import { Container } from '@/shared/components/shared';
-import { BlockCarousel } from '@/shared/components/shared/carousel/carousel';
-
-export default async function ServerCarousel() {
-  const carouselFiles = await getCarousel();
-  return (
-    <Container>
-      <BlockCarousel files={carouselFiles} />
-    </Container>
-  );
-}
